Actualiza versión de libGDX a 1.10.0 (latest)
Ref: https://libgdx.com/news/2021/04/gdx-1-10

+++ client/src/game/systems/camera/CameraSystem.java
@@ -58,7 +58,7 @@ public class CameraSystem extends BaseSystem {
         }
     }
 
-    public void zoom(int inout, float duration) {
+    public void zoom(float inout, float duration) {
         cameraZoomOrigin = camera.zoom;
         desiredZoom += inout * 0.025f;
         desiredZoom = MathUtils.clamp(desiredZoom, minZoom, maxZoom);
